requirements.txt
google-generativeai
websockets
requests
